// frontend/src/types/movie.ts
export type Movie = {
  movieId: number
  title: string
  year: number
  director: string
  rating: number
  category: string
  edited: boolean
  lentTo: string
  notes: string
}

//"movieId":1,"title":"Ace Ventura: Pet Detective","year":1994,"director":"Tom Shadyac",
//"rating":0,"category":"Comedy","edited":false,"lentTo":"","notes":"?"